requirements.txt
openai==1.35.7
pillow==10.3.0
pillow_heif==0.16.0
